MASTER-1830 fix isign tests
Made isign script a simpler shell. Doesn't have tests since
it uses the same interface, eliminating need for hardcorded
paths in tests.

Reorganized tests, now dependent on an IsignBaseTest
